+++ src/main/java/com/bewitchment/common/brew/GrowFlowersBrew.java
@@ -1,13 +1,14 @@
 package com.bewitchment.common.brew;
 
+import java.util.Random;
+
+import javax.annotation.Nullable;
+
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.init.Blocks;
 import net.minecraft.util.EnumFacing;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
-
-import javax.annotation.Nullable;
-import java.util.Random;
 
 /**
  * This class was created by Arekkuusu on 11/06/2017.
@@ -29,7 +30,7 @@ public class GrowFlowersBrew extends BlockHitBrew {
 	@Override
 	public void safeImpact(BlockPos pos, @Nullable EnumFacing side, World world, int amplifier) {
 		Random rand = world.rand;
-		int box = 1 + (int) ((float) amplifier / 2F);
+		int box = 1 + (int) (amplifier / 2F);
 
 		BlockPos posI = pos.add(box, box, box);
 		BlockPos posF = pos.add(-box, -box, -box);
